requirements.txt
aio-pika~=6.7.0
aiodns~=2.0.0
aiohttp~=3.6.2
aiomisc~=10.1.0
asyncpg~=0.21.0
boto3~=1.15.6
cchardet~=2.1.6
ciso8601~=2.1.3
envparse~=0.2.0
tortoise-orm~=0.16.16
ujson~=3.0.0
uvloop~=0.14.0
ffmpeg-python>=0.2.0,<1.0.0
